replaced all c-style casts with their c++ equivalents
git-svn-id: https://svn.code.sf.net/p/parity/code/trunk@135 26386e8f-9d3d-0410-aaca-840ac0c00a30

+++ parity.test/parity.binary.object.cpp
@@ -86,7 +86,7 @@ namespace parity
 				{
 					++countAllSyms;
 
-					if((unsigned int)sym->second.getSectionNumber() == it->first)
+					if(sym->second.getSectionNumber() == static_cast<short>(it->first))
 						utils::Log::verbose(" * SYM: Index: %5d, Name:, %s\n", sym->first, sym->second.getName().c_str());
 				}
 			}
